Improve performance
This was quite an adventure, it started with re-writing how entities,
particles, and bullets are created to make sure they don't require an
unbounded amount of memory.

The real issue was that a draw target was being re-created each frame,
and we were drawing the ward sprites to it. The pixelgl drawer was
caching each draw target as per
https://godoc.org/github.com/faiface/pixel#Drawer